+++ frontend/src/screens/VerificationScreen.js
@@ -15,24 +15,30 @@ const VerificationScreen = (props) => {
     const code = props.match.params.code;
 
     const dispatch = useDispatch();
-    const userRegister = useSelector(state => state.userRegister);
-    const {error} = userRegister;
+    const userVerify = useSelector(state => state.userVerify);
+    const {error, userInfo, loading} = userVerify;
     let history = useHistory();
+    const user = localStorage.getItem('userInfo');
 
 
     useEffect(() => {
-        dispatch(verify(code));
-        if(!error){
-            history.push("/admin");
+        async function verify_user() {
+            await dispatch(verify(code));
+            if(user){
+                history.push('/');
+                window.location.reload();
+            }
+            console.log(user);
         }
-    }, [history]);
+        verify_user();
+    }, [history, user]);
 
     return(
         <div className="body-verify">
             <Alert>
-                Подтверждение через почту
-                <Spinner animation="border"/>
-                {error}
+            {error ? error : "Подтверждение через почту"}                
+                {loading &&
+                <Spinner animation="border"/> }
             </Alert>
         </div>
     )